PostgresPythonGunicorn/Gunicorn/requirements.txt
psycopg2-binary==2.9.1
gunicorn==20.1.0
requests==2.25.1